feat: project structure change
