+++ package.json
@@ -2,9 +2,9 @@
     "name": "ws_task",
     "version": "1.0.0",
     "description": "Template for the WebSocket remote control task",
-    "main": "index.js",
+    "main": "dist/index.js",
     "scripts": {
-        "start": "node ./index.js",
+        "start": "tsc && node dist/index.js",
         "start:dev": "nodemon",
         "test": "echo \"Error: no test specified\" && exit 1"
     },
